# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IaProfileDifficulty =
  (typeof IaProfileDifficulty)[keyof typeof IaProfileDifficulty];

export const IaProfileDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface IaProfile {
  id: string;
  name: string;
  type: string;
  tagline: string;
  description: string;
  avatarUrl: string;
  coverUrl?: string;
  category: string;
  difficulty: IaProfileDifficulty;
  videoCount: number;
  missionCount: number;
  followerCount: number;
  createdAt: string;
}

export interface IaVideo {
  id: string;
  iaId: string;
  iaName: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  videoUrl: string;
  durationSeconds: number;
  views: number;
  publishedAt: string;
}

export type MissionStatus = (typeof MissionStatus)[keyof typeof MissionStatus];

export const MissionStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface Mission {
  id: string;
  iaId: string;
  order: number;
  title: string;
  description: string;
  content: string;
  requirements: string[];
  rewardPoints: number;
  status: MissionStatus;
}

export interface OverviewStats {
  totalIas: number;
  totalVideos: number;
  totalMissions: number;
  completedMissions: number;
  totalPoints: number;
  streakDays: number;
}

export type ActivityItemAction =
  (typeof ActivityItemAction)[keyof typeof ActivityItemAction];

export const ActivityItemAction = {
  started: "started",
  completed: "completed",
} as const;

export interface ActivityItem {
  id: string;
  iaId: string;
  iaName: string;
  iaAvatarUrl: string;
  missionTitle: string;
  action: ActivityItemAction;
  at: string;
}

export interface IaProgress {
  iaId: string;
  total: number;
  done: number;
  inProgress: number;
  percentComplete: number;
  earnedPoints: number;
}

export type ListIaProfilesParams = {
  search?: string;
  type?: string;
};

export type ListIaVideosParams = {
  iaId?: string;
};

export type ListIaMissionsParams = {
  iaId: string;
};

export type UpdateMissionStatusBodyStatus =
  (typeof UpdateMissionStatusBodyStatus)[keyof typeof UpdateMissionStatusBodyStatus];

export const UpdateMissionStatusBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateMissionStatusBody = {
  status: UpdateMissionStatusBodyStatus;
};