# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemAction } from "./activityItemAction";

export interface ActivityItem {
  id: string;
  iaId: string;
  iaName: string;
  iaAvatarUrl: string;
  missionTitle: string;
  action: ActivityItemAction;
  at: Date;
}