# lib/api-zod/src/generated/types/iaProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */
import type { IaProfileDifficulty } from "./iaProfileDifficulty";

export interface IaProfile {
  id: string;
  name: string;
  type: string;
  tagline: string;
  description: string;
  avatarUrl: string;
  coverUrl?: string;
  category: string;
  difficulty: IaProfileDifficulty;
  videoCount: number;
  missionCount: number;
  followerCount: number;
  createdAt: Date;
}