# lib/api-zod/src/generated/types/iaProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */

export interface IaProgress {
  iaId: string;
  total: number;
  done: number;
  inProgress: number;
  percentComplete: number;
  earnedPoints: number;
}