# lib/api-zod/src/generated/types/iaVideo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */

export interface IaVideo {
  id: string;
  iaId: string;
  iaName: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  videoUrl: string;
  durationSeconds: number;
  views: number;
  publishedAt: Date;
}