# lib/api-zod/src/generated/types/mission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */
import type { MissionStatus } from "./missionStatus";

export interface Mission {
  id: string;
  iaId: string;
  order: number;
  title: string;
  description: string;
  content: string;
  requirements: string[];
  rewardPoints: number;
  status: MissionStatus;
}