# lib/api-zod/src/generated/types/overviewStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telecurso de IAs API
 * OpenAPI spec version: 0.1.0
 */

export interface OverviewStats {
  totalIas: number;
  totalVideos: number;
  totalMissions: number;
  completedMissions: number;
  totalPoints: number;
  streakDays: number;
}